Move BSInvMarker back to object
